+++ src/colors/index.ts
@@ -1,6 +1,6 @@
 export * from './colorTypes';
 export * from './contrastCurve';
 export * from './materialDynamicColors';
-export * from './materialTheme';
+export * from './colorTheme';
 export * from './toneDeltaPair';
 export * from './variant';