+++ src/utils/extendTheme.ts
@@ -1,0 +1,5 @@
+import { Theme } from '../main';
+
+export interface ExtendTheme {
+  updateTheme: (theme: Theme) => Theme;
+}