+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ### Features
 
-* Updated MaterialFonts, typography styles, and Tailwind Material version ([aab7649](https://github.com/Udixio/tailwind-material/commit/aab7649810752cf18860cf96ac54476bed03d4e1))
+* Updated FontTheme, typography styles, and Tailwind Material version ([aab7649](https://github.com/Udixio/tailwind-material/commit/aab7649810752cf18860cf96ac54476bed03d4e1))
 
 # [1.0.0-beta.12](https://github.com/Udixio/tailwind-material/compare/v1.0.0-beta.11...v1.0.0-beta.12) (2023-11-09)